change node id from int to long.

+++ framework/src/main/java/edu/uci/eecs/wukong/framework/model/EndPoint.java
@@ -2,15 +2,15 @@ package edu.uci.eecs.wukong.framework.model;
 
 public class EndPoint {
 	/* Network address */
-	private int nodeId;
+	private long nodeId;
 	private byte portId;
 	
-	public EndPoint(int nodeId, byte portId) {
+	public EndPoint(long nodeId, byte portId) {
 		this.nodeId = nodeId;
 		this.portId = portId;
 	}
 	
-	public int getNodeId() {
+	public long getNodeId() {
 		return nodeId;
 	}
